source2topic/requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-pubsub==2.18.1
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==1.26.16
